RavenDB-17793 organizing code
